Broken down the stats to multiple smaller files
