+++ api/src/schema.gql
@@ -113,7 +113,7 @@ type Account {
   address: Bytes!
   balance: Decimal
   slowWallet: SlowWallet
-  reauthorized: Boolean!
+  initialized: Boolean
   movements(first: Int! = 10, after: String, order: OrderDirection! = ASC): PaginatedMovements!
 }
